Organize routes into separate files
- Moved homepage, about, search, and FAQ routes to their own files.
- Updated `server.js` to import and use route modules for homepage, about, search, and FAQ routes.
