+++ routes/search.js
@@ -1,0 +1,9 @@
+const express = require('express');
+const router = express.Router();
+
+// Route to render Search page
+router.get('/search', (req, res) => {
+  res.render('search');
+});
+
+module.exports = router;
